Moving PT tags to scenarios
